Detect whether a cpu has hypthreading
My hyperthreaded systems show a drastic performance drop during the
find_dupes stage when we use the # of processors reported by the system.
sysconf(_SC_NPROCESSORS_ONLN) can only give us the count of logical
processors, and there's no 'standard' way that I can see to get the physical
count.

So instead we parse /proc/cpuinfo and look for an Intel CPU. If we find
that, we check the flags field for 'ht'.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>